+++ src/main/java/com/gestion/personnel/controllers/EntrepriseController.java
@@ -4,7 +4,7 @@ import com.gestion.personnel.dto.DirectionDto;
 import com.gestion.personnel.dto.EntrepriseDto;
 import com.gestion.personnel.models.Direction;
 import com.gestion.personnel.models.Entreprise;
-import com.gestion.personnel.services.EntrepriseService;
+import com.gestion.personnel.services.EntrepriseServiceImpl;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -16,7 +16,7 @@ import java.util.List;
 public class EntrepriseController {
 
     @Autowired
-    private EntrepriseService entrepriseService;
+    private EntrepriseServiceImpl entrepriseService;
 
     // Lister toutes les entreprises
     @GetMapping("/lister")
