+++ src/main/java/com/gestion/personnel/dto/DirectionDto.java
@@ -1,10 +1,11 @@
 package com.gestion.personnel.dto;
 
 import lombok.AllArgsConstructor;
+import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
-
+@Builder
 @Getter
 @Setter
 @AllArgsConstructor
